requirements.txt
Authlib==0.15.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-OAuth==0.12
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.19.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.6
URLObject==2.4.3
Werkzeug==2.0.1
